Feature/#27 (#28)
* chore: gameContext import 개선

* chore: react 셋업 외

* feat: 초기 메인화면 UI, game start animation

* feat: 게임 시작 로직

* feat: 전반적인 flow

* feat: dashboard에 fps정보 추가, trail effect 잔버그 픽스

+++ src/game/EventDispatcher.ts
@@ -1,3 +1,5 @@
+import Game from '@game';
+
 import { StateComponent } from '@game/models/component';
 import { Entity } from '@game/models/entity';
 import { GameEvent } from '@game/models/event';
@@ -9,18 +11,30 @@ import EntityManager from './EntityManager';
  * state pattern과의 연결고리, 다른 방법?
  */
 export default class EventDispatcher {
+  private _game: Game;
   private _entityManager: EntityManager;
   private _stateComponents: StateComponent[];
 
-  constructor(entityManager: EntityManager, stateComponents: StateComponent[]) {
+  constructor(
+    game: Game,
+    entityManager: EntityManager,
+    stateComponents: StateComponent[]
+  ) {
+    this._game = game;
     this._entityManager = entityManager;
     this._stateComponents = stateComponents;
   }
 
   dispatch(event: GameEvent, entity: Entity) {
     if (event === GameEvent.Dead) {
-      this._entityManager.removeEntity(entity);
+      if (entity === this._entityManager.getPlayerEntity()) {
+        this._game.endRound();
+      } else {
+        this._entityManager.removeEntity(entity);
+      }
     }
+
+    console.info(`[EVENT: ${event}, ENTITY: ${entity}]`);
 
     const stateComponent = this._stateComponents[entity];
     stateComponent.state?.handleEvent(new CustomEvent(event));
